part 207 completed with bugs fixed

+++ database/migrations/2022_05_27_155411_rename_blog_post_tag_table_to_taggables.php
@@ -41,7 +41,7 @@ class RenameBlogPostTagTableToTaggables extends Migration
         Schema::disableForeignKeyConstraints();
 
         Schema::table('blog_post_tag', function (Blueprint $table) {
-            $table->unsignedInteger('blog_post_id')->index();
+            $table->unsignedBigInteger('blog_post_id')->index();
             $table->foreign('blog_post_id')->references('id')->on('blog_posts')
                 ->onDelete('cascade');
         });
